+++ src/Checkbox/Checkbox.driver.js
@@ -19,7 +19,11 @@ const checkboxDriverFactory = ({element, wrapper, component, eventTrigger}) => {
     focus: () => eventTrigger.focus(checkbox()),
     /** trigger blur on the element */
     blur: () => eventTrigger.blur(checkbox()),
-    hasFocusState: () => isClassExists(element, 'hasFocus'),
+    /**
+     * Focus related testing is done in e2e tests only.
+     * @deprecated
+     */
+    hasFocusState: () => element.getAttribute('data-focus'),
     isChecked: () => isClassExists(element, 'checked'),
     isDisabled: () => isClassExists(element, 'disabled'),
     isIndeterminate: () => isClassExists(element, 'indeterminate'),