+++ src/Checkbox/Checkbox.e2e.js
@@ -1,78 +1,89 @@
 import eyes from 'eyes.it';
 import {checkboxTestkitFactory, getStoryUrl, waitForVisibilityOf} from '../../testkit/protractor';
 import autoExampleDriver from 'wix-storybook-utils/AutoExampleDriver';
+import {runFocusTests} from '../common/Focusable/FocusableTestsE2E';
+
+const NO_DESCRIPTION = '';
 
 describe('Checkbox', () => {
   const storyUrl = getStoryUrl('4. Selection', '4.2 Checkbox');
   const checkboxDriver = checkboxTestkitFactory({dataHook: 'storybook-checkbox'});
 
-  const waitForCheckbox = () => waitForVisibilityOf(checkboxDriver.element(), 'Cannot find Checkbox');
-  const clickTab = () => browser.actions().sendKeys(protractor.Key.TAB).perform();
 
-  beforeAll(async () => {
-    await browser.get(storyUrl);
-  });
-
-  beforeEach(async () => {
-    await autoExampleDriver.reset();
-    await browser.$('body').click();
-    await waitForCheckbox();
-  });
-
-  eyes.it('should have default props', async () => {
-    expect(checkboxDriver.hasError()).toBe(false, 'hasError');
-    expect(checkboxDriver.isChecked()).toBe(false, 'isChecked');
-    expect(checkboxDriver.isFocused()).toBe(false, 'isFocused');
-    expect(checkboxDriver.isDisabled()).toBe(false, 'isDisabled');
-  });
-
-  eyes.it('should set checked state when clicked', async () => {
-    expect(checkboxDriver.isChecked()).toBe(false);
-    await checkboxDriver.click();
-    expect(checkboxDriver.isChecked()).toBe(true);
-  });
-
-  eyes.it('should show focused styles', async () => {
-    expect(checkboxDriver.isFocused()).toBe(false);
-    await clickTab();
-    expect(checkboxDriver.isFocused()).toBe(true);
-  });
-
-  describe('has error', () => {
+  describe(NO_DESCRIPTION, () => {
+    const waitForCheckbox = () => waitForVisibilityOf(checkboxDriver.element(), 'Cannot find Checkbox');
+    const clickTab = () => browser.actions().sendKeys(protractor.Key.TAB).perform();
 
     beforeEach(async () => {
-      await autoExampleDriver.setProps({hasError: true});
+      // TODO: We do browser.get() before EACH test in order to reset the focus.
+      // implmement a generic solution in AutoExampleDriver that will do
+      // propper reset of the focus, so we don't have to get the page,
+      // and thus the test will run faster.
+      await browser.get(storyUrl);
+
+      // No need for reset as long as we do browser.get() before each test.
+      // await autoExampleDriver.reset();
+      await waitForCheckbox();
     });
 
-    eyes.it('should show error styles', async () => {
-      expect(checkboxDriver.hasError()).toBe(true);
+    eyes.it('should have default props', async () => {
+      expect(checkboxDriver.hasError()).toBe(false, 'hasError');
+      expect(checkboxDriver.isChecked()).toBe(false, 'isChecked');
+      expect(checkboxDriver.isFocused()).toBe(false, 'isFocused');
+      expect(checkboxDriver.isDisabled()).toBe(false, 'isDisabled');
+    });
+
+    eyes.it('should set checked state when clicked', async () => {
+      expect(checkboxDriver.isChecked()).toBe(false);
+      await checkboxDriver.click();
+      expect(checkboxDriver.isChecked()).toBe(true);
     });
 
     eyes.it('should show focused styles', async () => {
-      expect(checkboxDriver.hasError()).toBe(true);
       expect(checkboxDriver.isFocused()).toBe(false);
       await clickTab();
       expect(checkboxDriver.isFocused()).toBe(true);
     });
+
+    describe('has error', () => {
+
+      beforeEach(async () => {
+        await autoExampleDriver.setProps({hasError: true});
+      });
+
+      eyes.it('should show error styles', async () => {
+        expect(checkboxDriver.hasError()).toBe(true);
+      });
+
+      eyes.it('should show focused styles', async () => {
+        expect(checkboxDriver.hasError()).toBe(true);
+        expect(checkboxDriver.isFocused()).toBe(false);
+        await clickTab();
+        expect(checkboxDriver.isFocused()).toBe(true);
+      });
+    });
+
+    describe('is disabled', () => {
+
+      beforeEach(async () => {
+        await autoExampleDriver.setProps({disabled: true});
+      });
+
+      eyes.it('should be disabled', async () => {
+        expect(checkboxDriver.isDisabled()).toBe(true);
+      });
+
+      eyes.it('should not be focusable', async () => {
+        expect(checkboxDriver.isDisabled()).toBe(true);
+        expect(checkboxDriver.isFocused()).toBe(false);
+        await clickTab();
+        expect(checkboxDriver.isFocused()).toBe(false);
+      });
+    });
   });
 
-  describe('is disabled', () => {
-
-    beforeEach(async () => {
-      await autoExampleDriver.setProps({disabled: true});
-    });
-
-    eyes.it('should be disabled', async () => {
-      expect(checkboxDriver.isDisabled()).toBe(true);
-    });
-
-    eyes.it('should not be focusable', async () => {
-      expect(checkboxDriver.isDisabled()).toBe(true);
-      expect(checkboxDriver.isFocused()).toBe(false);
-      await clickTab();
-      expect(checkboxDriver.isFocused()).toBe(false);
-    });
-
+  describe('Generic', () => {
+    runFocusTests(checkboxDriver, storyUrl);
   });
 });