+++ src/RadioGroup/RadioButton/RadioButton.protractor.driver.js
@@ -1,0 +1,21 @@
+import focusableDriverFactory from '../../common/Focusable/Focusable.protractor.driver';
+import {mergeDrivers} from '../../test-common';
+
+const buttonDriverFactory = element => {
+  const getRadio = () => element.$(`[data-hook="radiobutton-radio"]`);
+  const getTextChildren = () => element.$(`[data-hook="radiobutton-children"]`);
+
+  const focusableDriver = focusableDriverFactory({
+    rootElement: element,
+    nativeFocusableElement: element,
+    clickableElements: [getRadio, getTextChildren]
+  });
+
+  const publicDriver = {
+    // Empty driver
+  };
+
+  return mergeDrivers(publicDriver, focusableDriver);
+};
+
+export default buttonDriverFactory;