+++ src/RichTextArea/RichTextAreaButton.js
@@ -10,7 +10,7 @@ import OrderedList from './../Icons/dist/components/OrderedList';
 import Link from './../Icons/dist/components/Link';
 import Image from './../Icons/dist/components/Image';
 import styles from './RichTextAreaButton.scss';
-
+import {withFocusable, focusableStates} from '../common/Focusable';
 const buttons = {
   bold: {
     icon: Bold,
@@ -51,11 +51,6 @@ const buttons = {
 
 class RichTextAreaButton extends Component {
 
-  constructor(props) {
-    super(props);
-    this.state = {isFocused: false};
-  }
-
   handleMouseDown = event => {
     event.preventDefault();
     if (!this.props.disabled) {
@@ -84,11 +79,11 @@ class RichTextAreaButton extends Component {
         <button
           type="button"
           className={className}
-          onMouseDown={this.handleMouseDown}
           data-hook={`rich-text-area-button-${type}`}
-          data-focused={this.state.isFocused}
-          onFocus={this.handleFocus}
-          onBlur={this.handleBlur}
+          {...focusableStates(this.props)}
+          onFocus={this.props.focusableOnFocus} // eslint-disable-line react/prop-types
+          onBlur={this.props.focusableOnBlur} // eslint-disable-line react/prop-types
+          onMouseDown={this.handleMouseDown} // override focusable's onMouseDown
           >
           <span className={styles.wrapper}>
             {this.renderIcon()}
@@ -103,12 +98,6 @@ class RichTextAreaButton extends Component {
     return <Icon size={`${size}px`}/>;
   }
 
-  handleFocus = () =>
-    !this.state.isFocused && this.setState({isFocused: true});
-
-  handleBlur = () =>
-    this.state.isFocused && this.setState({isFocused: false});
-
 }
 
 RichTextAreaButton.propTypes = {
@@ -119,4 +108,4 @@ RichTextAreaButton.propTypes = {
   disabled: PropTypes.bool
 };
 
-export default RichTextAreaButton;
+export default withFocusable(RichTextAreaButton);